Fix loading of saved graphs
Loading of saved graphs did not make copies of the curve objects and
thus any quick zoom/smooth/expo modifications of the curve affected
the saved template (not the one saved in the preferences, but that in
memory).

Additional effect was that for saved graphs using xxx[all] curves
all child curves would use the same curve object which makes it
impossible to zoom/smooth/expo individually.