Hanning window applied to input data (fftInput, not fftOutput)
A Hanning window is a taper that "tapers" the edges of data segments before performing fft, so the Fourier output does not contain artifacts from the sharp edges at the start and end of data segments. Here, it was incorrectly applied to the output data (the Fourier data). The result was streak artifacts along the frequency domain of the "Freq. vs Throttle" plots in BBE. Here, I simple moved the Hanning window portion of the code before fft was computed and applied it to Input data ('fftInput') instead of output data ('fftOutput').